Add files via upload
In these days it is common to control electrical appliances and gadgets using switches and infrared remote. Now it is the time for new control system to replace the existing one.A gesture is a type of nonverbal communication in which the hands or other different parts of the body are moved. Gestures are read by a camera. The human body’s movements are read and transmits the information to a computer employs gestures as a control input for devices or applications. In this pandemic times,this promotes safety. It will enable the use of contact-less equipment in public spaces, assuring the safety of those who use it. Along with these benefits, the primary purpose of home automation is to give users both comfort and peace of mind with safety precautions. Home appliances like fans, lights, etc will be controlled by the Hand gestures without any physical activities which is most essential in this pandemic situations and future.